refactor: Extract "Hide fullscreen ads" into `shared-youtube` path
