+++ patches/src/main/kotlin/app/morphe/patches/shared/ad/HideFullscreenAdsPatch.kt
@@ -22,8 +22,9 @@ import com.android.tools.smali.dexlib2.iface.instruction.TwoRegisterInstruction
 import com.android.tools.smali.dexlib2.iface.reference.FieldReference
 import com.android.tools.smali.dexlib2.iface.reference.MethodReference
 
+private const val FULLSCREEN_ADS_CLASS = "Lapp/morphe/extension/shared/patches/HideFullscreenAdsPatch;"
+
 internal fun hideFullscreenAdsPatch(
-    extensionClass: String,
     preferenceScreen: BasePreferenceScreen.Screen
 ) = bytecodePatch(
     description = "Adds an option to hide fullscreen premium popup ads."
@@ -50,8 +51,8 @@ internal fun hideFullscreenAdsPatch(
                     insertIndex,
                     """
                         move-object/from16 v$freeRegister, p1
-                        invoke-static { v$insertRegister, v$freeRegister }, $extensionClass->closeFullscreenAd(Ljava/lang/Object;[B)V
-                    """
+                        invoke-static { v$insertRegister, v$freeRegister }, $FULLSCREEN_ADS_CLASS->closeFullscreenAd(Ljava/lang/Object;[B)V
+                        """
                 )
             }
         }
